support page other sections added but responsive not done
