+++ src/Pages/Support/Components/FAQSection/FAQSection.jsx
