+++ src/Pages/Support/Components/FAQSection/lottie.json
