--- src/Pages/SupportSection/Components/SupportBanner/SupportBanner.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const SupportBanner = () => {
-  return (
-    <div>
-      <div className="text-white text-center text-4xl">Support Banner Upcoming</div>
-    </div>
-  );
-};
-
-export default SupportBanner;
